move modality-handler to its own file
