+++ shells/lib/source/arcslib.js
@@ -9,7 +9,8 @@
  */
 
 import {Arc} from '../../../build/runtime/arc.js';
-import {Modality, ModalityHandler} from '../../../build/runtime/modality.js';
+import {Modality} from '../../../build/runtime/modality.js';
+import {ModalityHandler} from '../../../build/runtime/modality-handler.js';
 import {Planificator} from '../../../build/runtime/plan/planificator.js';
 import {Suggestion} from '../../../build/runtime/plan/suggestion.js';
 import {SlotComposer} from '../../../build/runtime/slot-composer.js';
